requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# AI APIs
google-generativeai==0.3.2

# Environment and Security
python-dotenv==1.0.0
slowapi==0.1.9

# Utilities
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Firebase for cross-device session sync
firebase-admin==6.4.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==24.1.1
flake8==7.0.0
